server/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
google-cloud-storage==2.14.0
ray[client,default]==2.44.1 